# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GitHub Migrator API",
    "description": "API for managing repository discovery, profiling, batch organization, and migration execution between GitHub instances.",
    "version": "1.2.0",
    "contact": {
      "name": "GitHub Migrator"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/sources/{id}": {
      "get": {
        "tags": ["sources"],
        "summary": "Get source",
        "description": "Get a single source by ID",
        "operationId": "getSource",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Source details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Source"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": ["sources"],
        "summary": "Update source",
        "description": "Update an existing source (admin only)",
        "operationId": "updateSource",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Source updated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["sources"],
        "summary": "Delete source",
        "description": "Delete a source (admin only). Fails if repositories are associated.",
        "operationId": "deleteSource",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Source deleted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Source has associated repositories"
          }
        }
      }
    },
...
    "/api/v1/repositories/{fullName}": {
      "get": {
        "tags": ["repositories"],
        "summary": "Get repository details",
        "description": "Get detailed information about a specific repository including migration history",
        "operationId": "getRepository",
        "parameters": [
          {
            "$ref": "#/components/parameters/fullName"
          }
        ],
        "responses": {
          "200": {
            "description": "Repository details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "repository": {
                      "$ref": "#/components/schemas/Repository"
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MigrationHistory"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["repositories"],
        "summary": "Update repository",
        "description": "Update repository metadata such as batch assignment and priority",
        "operationId": "updateRepository",
        "parameters": [
          {
            "$ref": "#/components/parameters/fullName"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "batch_id": {
                    "type": "integer",
                    "format": "int64",
                    "nullable": true
                  },
                  "priority": {
                    "type": "integer"
                  },
                  "destination_full_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repository updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Repository"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Repository": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "full_name": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "source_url": {
            "type": "string"
          },
          "total_size": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "has_lfs": {
            "type": "boolean"
          },
          "has_submodules": {
            "type": "boolean"
          },
          "has_large_files": {
            "type": "boolean"
          },
          "has_actions": {
            "type": "boolean"
          },
          "has_wiki": {
            "type": "boolean"
          },
          "has_pages": {
            "type": "boolean"
          },
          "has_discussions": {
            "type": "boolean"
          },
          "has_projects": {
            "type": "boolean"
          },
          "is_archived": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "batch_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "priority": {
            "type": "integer"
          },
          "complexity_score": {
            "type": "integer"
          },
          "destination_url": {
            "type": "string",
            "nullable": true
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "migrated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "MigrationHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "repository_id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_message": {
            "type": "string",
            "nullable": true
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "duration_seconds": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
      "Source": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["github", "azuredevops"]
          },
          "base_url": {
            "type": "string"
          },
          "has_app_auth": {
            "type": "boolean"
          },
          "is_active": {
            "type": "boolean"
          },
          "repository_count": {
            "type": "integer"
          },
          "last_sync_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "masked_token": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "parameters": {
      "fullName": {
        "name": "fullName",
        "in": "path",
        "required": true,
        "description": "Repository full name (org/repo)",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}